requirements/base.txt
django==4.2.10
psycopg2-binary==2.9.9
python-environ==0.4.54
django-cors-headers==4.3.1
django-jazzmin==2.6.1
django-solo==2.2.0
drf-yasg==1.21.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.1
django-modeltranslation==0.18.11
Pillow==10.2.0
django-ckeditor==6.7.1
django-resized==1.0.2
python-telegram-bot==13.15  # last sync version
django-jsonform==2.22.0
django-tinymce==4.1.0
django-nested-inline==0.4.6
